process_data/requirements.txt
pandas
psycopg2-binary==2.9.6
SQLAlchemy==2.0.20
